moved the dungeon base, removed gold dungeon dependency to base
